+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add test-related sandbox configurations and runtime fixes
 - Set the required `extracted` and `collected` attributes for custom test classpath configurations
 - Set the repository name and fix the `action` application in `createCustomPluginRepository`
+- Refactor build service registration to use `registerClassLoaderScopedBuildService` to avoid issues caused because of different classpath in the project. JetBrains/intellij-platform-gradle-plugin#1919
 
 ## [2.5.0] - 2025-04-01